+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 1.4.0 / ####-##-##
 ==================
 
+  * Plugin system for Wendigo browser
   * Mock assert postBody
   * Logs parse js objects correctly in console module
   * Minor improvements in mock handling
+  * Minor code improvements on browserFactory and Modules
 
 1.3.0 / 2018-10-01
 ==================